socketio-server/requirements.txt
python-socketio==5.11.1
tornado==6.4
aiohttp==3.9.3